update promotionId to long

+++ WalletService.Application/Requests/BonusAccounts/Commands/AddBonusCreditTransaction/AddBonusCreditTransactionCommand.cs
@@ -4,6 +4,6 @@ namespace WalletService.Application.Requests.BonusAccounts.Commands.AddBonusCred
 
 public record AddBonusCreditTransactionCommand(Guid AccountId, string AccountType, string TransactionNo,
     string TransactionReference, decimal Amount, string ModeOfTransaction, string? Notes,
-    int PromotionId, DateTime DateStarted, DateTime DateExpired) : IRequest<Unit>
+    long PromotionId, DateTime DateStarted, DateTime DateExpired) : IRequest<Unit>
 {
 }
