+++ WalletService.Application/Requests/BonusAccounts/Commands/AddBonusDebitTransaction/AddBonusDebitTransactionCommand.cs
@@ -4,4 +4,4 @@ namespace WalletService.Application.Requests.BonusAccounts.Commands.AddBonusDebi
 
 public record AddBonusDebitTransactionCommand(Guid AccountId, string AccountType, string TransactionNo,
     string TransactionReference, string ModeOfTransaction, decimal Amount, string Notes,
-    int PromotionId, DateTime DateStarted, DateTime DateExpired) : IRequest<Unit>;
+    long PromotionId, DateTime DateStarted, DateTime DateExpired) : IRequest<Unit>;
