+++ WalletService.Application/Requests/BonusAccounts/Commands/CreateBonusAccount/CreateBonusAccountCommand.cs
@@ -4,5 +4,5 @@ namespace WalletService.Application.Requests.BonusAccounts.Commands.CreateBonusA
 
 public record CreateBonusAccountCommand(Guid AccountId, string AccountType, string TransactionNo, string TransactionReference,
     decimal Amount, string ModeOfTransaction, string Notes,
-    int PromotionId, DateTime DateStarted, DateTime DateExpired) : IRequest<Unit>;
+    long PromotionId, DateTime DateStarted, DateTime DateExpired) : IRequest<Unit>;
