+++ WalletService.Domain/Entities/BonusWalletTransaction.cs
@@ -9,7 +9,7 @@ public class BonusWalletTransaction
     public Guid AccountId { get; set; }
     public string TransactionNo { get; set; }
     public TransactionType TransactionType { get; set; } = TransactionType.Debit;
-    public int PromotionId { get; set; }
+    public long PromotionId { get; set; }
     public DateTime DateStarted { get; set; } = DateTime.Now;
     public DateTime DateExpired { get; set; }
     public string TransactionReference { get; set; }
